pyproject.toml
[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "EatCookiesBot"
version = "0.1.0"
description = ""
authors = ["Alexey Vasilenko <avenior@ya.ru>"]

[tool.poetry.dependencies]
python = "^3.8"
aiogram = "^2.18"
ujson = "^5.1.0"
aiohttp = {extras = ["speedups"], version = "^3.8.1"}
loguru = "^0.5.3"
python-dotenv = "^0.19.2"
SQLAlchemy = "^1.4.29"
psycopg2 = "^2.9.3"
dateutils = "^0.6.12"

[tool.poetry.dev-dependencies]
nitpick = "^0.29.0"
pre-commit = "^2.6.0"
pre-commit-hooks = "^4.1.0"
wemake-python-styleguide = "^0.16.0"
ipdb = "^0.13.8"
ipython = "^7.23.1"

# Code Formatters
autoflake = "^1.3.1"
autopep8 = "^1.6.0"
docformatter = "^1.3.1"
isort = "^5.10.1"
language-formatters-pre-commit-hooks = "^2.2.0"

# (Static) Code Analyzers
darglint = "^1.5"
flake8 = "^4.0.1"
flake8-bandit = "^2.1"
flake8-broken-line = "^0.4.0"
flake8-bugbear = "^21.11.29"
flake8-comprehensions = "^3.2.0"
flake8-debugger = "^4.0.0"
flake8-docstrings = "^1.5"
flake8-eradicate = "^1.2.0"
flake8-fixme = "^1.1.1"
flake8-isort = "^4.1.1"
flake8-logging-format = "^0.6.0"
flake8-pytest-style = "^1.3"
flake8-quotes = "^3.3.1"
flake8-rst-docstrings = "^0.2.5"
flake8-string-format = "^0.3.0"
flake8_commas = "^2.0"
mypy = "^0.931"
pycodestyle = "^2.6"
pydocstyle = "^6.1.1"
sqlalchemy-stubs = "^0.4"

# Editor integration
python-lsp-server = "^1.1.0"
pyls-isort = "^0.2.2"
pyls-flake8 = "^0.4.0"
mypy-ls = "^0.5.1"

# Test Suite
pytest = "^6.2.5"
pytest-asyncio = "^0.16.0"
pytest-cov = "^3.0.0"
hypothesis = "^6.1.1"
oitnb = "^0.2.2"
types-python-dateutil = "^2.8.6"
